move processRpcComment to outstanding folder
